add mockpubsub to help tests
