+++ service/src/test/java/org/databiosphere/workspacedataservice/pubsub/MockPubSub.java
@@ -1,0 +1,8 @@
+package org.databiosphere.workspacedataservice.pubsub;
+
+public class MockPubSub implements PubSub {
+
+  public void publish(String message) {
+    // no-op
+  }
+}
